Create currentDate module using luxon
