+++ modules/book.js
@@ -1,11 +1,10 @@
-export class Book {
+export default class Book {
     static id = -1;
-  
+
     constructor(title, author) {
       this.title = title;
       this.author = author;
       Book.id += 1;
       this.idx = Book.id;
     }
-  }
-  
+}
