+++ modules/bookCollection.js
@@ -1,24 +1,24 @@
-import {Book} from './book.js';
+import Book from './book.js';
 
-export class BookCollection {
-    constructor() {
-      this.books = [];
-    }
-  
-    add(title, author) {
-      const newBook = new Book(title, author);
-      this.books.push(newBook);
-    }
-  
-    remove(id) {
-      this.books = this.books.filter((book) => book.idx !== parseInt(id, 10));
-    }
-  
-    getBooks() {
-      return this.books;
-    }
-  
-    setBooks(books) {
-      this.books = books;
-    }
+export default class BookCollection {
+  constructor() {
+    this.books = [];
   }
+
+  add(title, author) {
+    const newBook = new Book(title, author);
+    this.books.push(newBook);
+  }
+
+  remove(id) {
+    this.books = this.books.filter((book) => book.idx !== parseInt(id, 10));
+  }
+
+  getBooks() {
+    return this.books;
+  }
+
+  setBooks(books) {
+    this.books = books;
+  }
+}
