+++ modules/currentDate.js
@@ -1,0 +1,6 @@
+import DateTime from "../node_modules/luxon/src/datetime.js";
+export function displayCurrentDate (dateContainer){
+setInterval(() => {
+    dateContainer.innerHTML = DateTime.now().toLocaleString(DateTime.DATETIME_MED_WITH_SECONDS);
+  },1000);
+}
