+++ package.json
@@ -244,6 +244,7 @@
     "log-symbols": "^4.1.0",
     "lowercase-keys": "^2.0.0",
     "lru-cache": "^6.0.0",
+    "luxon": "^2.3.0",
     "make-dir": "^3.1.0",
     "make-fetch-happen": "^9.1.0",
     "map-age-cleaner": "^0.1.3",
